# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: keel API
  version: 1.0.0
  description: |
    Contract-first API (anayasa K3, ADR-006, ADR-024). The backend implements this file; the mobile client's types are
    generated from it, never hand-written. Rules the contract keeps, each checked by backend ContractTests:
    - The Decision is the engine's Decision field for field (U3). Actions are a tagged union on `type`.
    - Every calorie estimate is a range, `KcalRange` / `KcalBalance` (U5). A target and a decision's step may be one
      number (U5); a device's own reading is passed through as given.
    - The engine's fat estimate never appears, in any form (U4). Error messages never carry health data (V3).
    Conventions: everything but health and sign-in is under /v1 and needs a session. Records the phone creates
    offline carry a `clientId` (a UUID the phone makes): sending one twice stores it once and answers 200 with the
    stored record instead of 201. Days are `date` (the user's local calendar day); moments are `date-time` with offset.
    Lists page backwards: `limit` and `before` (the cursor of the last item seen).
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - session: []
...
paths:
...
  /v1/weigh-ins:
    get:
      tags: [measurement]
      operationId: listWeighIns
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Weigh-ins in the range, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeighIn"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [measurement]
      operationId: recordWeighIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewWeighIn"
      responses:
        "201":
          $ref: "#/components/responses/WeighInCreated"
        "200":
          $ref: "#/components/responses/WeighInCreated"
        default:
          $ref: "#/components/responses/Error"
  /v1/weigh-ins/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    delete:
      tags: [measurement]
      operationId: deleteWeighIn
      responses:
        "204":
          description: Deleted
        default:
          $ref: "#/components/responses/Error"
  /v1/weight-trend:
    get:
      tags: [measurement]
      operationId: getWeightTrend
      summary: The trend weight per day (7-day average, the engine's WeightTrend)
      description: |
        One point per day, up to today in the user's time zone, whose 7 days hold at least one weigh-in (the day's first
        counts). Needs the profile (NOT_FOUND without one). At most 400 days per request.
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: One point per day that has a trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"
        default:
          $ref: "#/components/responses/Error"
  /v1/waist-measurements:
    get:
      tags: [measurement]
      operationId: listWaistMeasurements
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Waist measurements in the range, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WaistMeasurement"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [measurement]
      operationId: recordWaistMeasurement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewWaistMeasurement"
      responses:
        "201":
          $ref: "#/components/responses/WaistMeasurementCreated"
        "200":
          $ref: "#/components/responses/WaistMeasurementCreated"
        default:
          $ref: "#/components/responses/Error"
  /v1/photo-checks:
    post:
      tags: [measurement]
      operationId: recordPhotoCheck
      summary: What the phone concluded from a progress photo; the photo itself never leaves the phone (V1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPhotoCheck"
      responses:
        "201":
          $ref: "#/components/responses/PhotoCheckCreated"
        "200":
          $ref: "#/components/responses/PhotoCheckCreated"
        default:
          $ref: "#/components/responses/Error"
  /v1/activity-days:
    put:
      tags: [measurement]
      operationId: putActivityDay
      summary: A day's steps, sleep and active energy from Apple Health (replaces that day's values)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityDay"
      responses:
        "200":
          description: Stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityDay"
        default:
          $ref: "#/components/responses/Error"

  # ── nutrition (K-208, K-209, ADR-008) ──────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The access token of a Session (K-203).
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    From:
      name: from
      in: query
      required: true
      description: First day, inclusive (local day). A range is at most 400 days.
      schema:
        type: string
        format: date
    To:
      name: to
      in: query
      required: true
      description: Last day, inclusive
      schema:
        type: string
        format: date
...
  responses:
    Error:
      description: Any error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    WeighInCreated:
      description: Stored (201), or already stored with this clientId (200)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WeighIn"
    WaistMeasurementCreated:
      description: Stored (201), or already stored with this clientId (200)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WaistMeasurement"
    PhotoCheckCreated:
      description: Stored (201), or already stored with this clientId (200)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PhotoCheck"
...
  schemas:
...
    Error:
      type: object
      description: Shared error body. Messages never contain health data (anayasa V3).
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable code, e.g. VALIDATION_FAILED, NOT_FOUND, CONSENT_REQUIRED.
        message:
          type: string
          description: English, safe to show to the user.
...
    ClientId:
      type: string
      format: uuid
      description: Made by the phone when the record is created; the same record sent again is stored once.

    # ── auth ─────────────────────────────────────────────────────────────────────────────────────────────────
...
    MeasurementSource:
      type: string
      enum: [MANUAL, APPLE_HEALTH, IMPORT]
    NewWeighIn:
      type: object
      required: [clientId, measuredAt, kg, source]
      properties:
        clientId:
          $ref: "#/components/schemas/ClientId"
        measuredAt:
          type: string
          format: date-time
        kg:
          type: number
          exclusiveMinimum: 0
          maximum: 500
          description: At most 2 decimals (the store keeps 2; a finer value is a validation error, not a silent rounding).
        source:
          $ref: "#/components/schemas/MeasurementSource"
    WeighIn:
      allOf:
        - $ref: "#/components/schemas/NewWeighIn"
        - type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
    TrendPoint:
      type: object
      required: [day, kg]
      properties:
        day:
          type: string
          format: date
        kg:
          type: number
    NewWaistMeasurement:
      type: object
      required: [clientId, measuredOn, cm]
      properties:
        clientId:
          $ref: "#/components/schemas/ClientId"
        measuredOn:
          type: string
          format: date
        cm:
          type: number
          exclusiveMinimum: 0
          maximum: 300
          description: At most 1 decimal.
    WaistMeasurement:
      allOf:
        - $ref: "#/components/schemas/NewWaistMeasurement"
        - type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
    Look:
      type: string
      description: The photo compared with the last one, judged on the phone.
      enum: [BETTER, SAME, WORSE]
    NewPhotoCheck:
      type: object
      description: Derived on the phone; no image, no measurement of the body leaves it (V1).
      required: [clientId, takenOn, look]
      properties:
        clientId:
          $ref: "#/components/schemas/ClientId"
        takenOn:
          type: string
          format: date
        look:
          $ref: "#/components/schemas/Look"
    PhotoCheck:
      allOf:
        - $ref: "#/components/schemas/NewPhotoCheck"
        - type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
    ActivityDay:
      type: object
      required: [day]
      properties:
        day:
          type: string
          format: date
        steps:
          type: integer
          minimum: 0
        sleepMinutes:
          type: integer
          minimum: 0
        activeEnergyKcal:
          type: integer
          minimum: 0
          description: Apple Health's own figure, passed through as given; any range around it is made on the server.

    # ── nutrition ────────────────────────────────────────────────────────────────────────────────────────────
...